[TDD]기본과제 / by.yeonjin / 4가지 기본 기능 구현 변경
