--- src/test/java/io/hhplus/tdd/point/PointControllerTest.java
@@ -1,11 +1,0 @@
-package io.hhplus.tdd.point;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-@SpringBootTest
-public class PointControllerTest {
-
-}
